HARMONY-1652: Add test for workflows with
